# src/main/resources/swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Api Documentation",
    "description": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Inferred Url"
    }
  ],
  "tags": [
    {
      "name": "bank-account-controller",
      "description": "Bank Account Controller"
    },
    {
      "name": "basic-error-controller",
      "description": "Basic Error Controller"
    }
  ],
  "paths": {
    "/bank/api/deposit": {
      "post": {
        "tags": [
          "bank-account-controller"
        ],
        "summary": "deposit",
        "operationId": "depositUsingPOST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEntity"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/bank/api/{customerId}/bankaccount/activate": {
      "post": {
        "tags": [
          "bank-account-controller"
        ],
        "summary": "activateAccount",
        "operationId": "activateAccountUsingPOST",
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "description": "customerId",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IbanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEntity"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/bank/api/{customerId}/bankaccount/create": {
      "post": {
        "tags": [
          "bank-account-controller"
        ],
        "summary": "createAccount",
        "operationId": "createAccountUsingPOST",
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "description": "customerId",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEntity"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/bank/api/{customerId}/bankaccount/deactivate": {
      "post": {
        "tags": [
          "bank-account-controller"
        ],
        "summary": "deactivateAccount",
        "operationId": "deactivateAccountUsingPOST",
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "description": "customerId",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IbanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEntity"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/bank/api/{customerId}/bankaccounts": {
      "get": {
        "tags": [
          "bank-account-controller"
        ],
        "summary": "getBankAccounts",
        "operationId": "getBankAccountsUsingGET",
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "description": "customerId",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "description": "filter",
            "required": false,
            "style": "form",
            "schema": {
              "type": "string",
              "enum": [
                "CHECKINGS",
                "PRIVATE_LOAN",
                "SAVINGS"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEntity"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/bank/api/{customerId}/transfer": {
      "post": {
        "tags": [
          "bank-account-controller"
        ],
        "summary": "transfer",
        "operationId": "transferUsingPOST",
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "description": "customerId",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEntity"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/bank/api/{customerId}/{iban}/balance": {
      "get": {
        "tags": [
          "bank-account-controller"
        ],
        "summary": "checkBalance",
        "operationId": "checkBalanceUsingGET",
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "description": "customerId",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "iban",
            "in": "path",
            "description": "iban",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEntity"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/bank/api/{iban}/transactions": {
      "get": {
        "tags": [
          "bank-account-controller"
        ],
        "summary": "transactions",
        "operationId": "transactionsUsingGET",
        "parameters": [
          {
            "name": "iban",
            "in": "path",
            "description": "iban",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEntity"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/bank/error": {
      "get": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "error",
        "operationId": "errorUsingGET",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "error",
        "operationId": "errorUsingPUT",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "error",
        "operationId": "errorUsingPOST",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "error",
        "operationId": "errorUsingDELETE",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "options": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "error",
        "operationId": "errorUsingOPTIONS",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "head": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "error",
        "operationId": "errorUsingHEAD",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "patch": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "error",
        "operationId": "errorUsingPATCH",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "trace": {
        "tags": [
          "basic-error-controller"
        ],
        "summary": "error",
        "operationId": "errorUsingTRACE",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateAccountDto": {
        "title": "CreateAccountDto",
        "type": "object",
        "properties": {
          "accountType": {
            "type": "string",
            "enum": [
              "CHECKINGS",
              "PRIVATE_LOAN",
              "SAVINGS"
            ]
          }
        }
      },
      "DepositRequestDto": {
        "title": "DepositRequestDto",
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "format": "double"
          },
          "iban": {
            "type": "string"
          }
        }
      },
      "IbanDto": {
        "title": "IbanDto",
        "type": "object",
        "properties": {
          "iban": {
            "type": "string"
          }
        }
      },
      "ModelAndView": {
        "title": "ModelAndView",
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "model": {
            "type": "object"
          },
          "modelMap": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          },
          "reference": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACCEPTED",
              "ALREADY_REPORTED",
              "BAD_GATEWAY",
              "BAD_REQUEST",
              "BANDWIDTH_LIMIT_EXCEEDED",
              "CHECKPOINT",
              "CONFLICT",
              "CONTINUE",
              "CREATED",
              "DESTINATION_LOCKED",
              "EXPECTATION_FAILED",
              "FAILED_DEPENDENCY",
              "FORBIDDEN",
              "FOUND",
              "GATEWAY_TIMEOUT",
              "GONE",
              "HTTP_VERSION_NOT_SUPPORTED",
              "IM_USED",
              "INSUFFICIENT_SPACE_ON_RESOURCE",
              "INSUFFICIENT_STORAGE",
              "INTERNAL_SERVER_ERROR",
              "I_AM_A_TEAPOT",
              "LENGTH_REQUIRED",
              "LOCKED",
              "LOOP_DETECTED",
              "METHOD_FAILURE",
              "METHOD_NOT_ALLOWED",
              "MOVED_PERMANENTLY",
              "MOVED_TEMPORARILY",
              "MULTIPLE_CHOICES",
              "MULTI_STATUS",
              "NETWORK_AUTHENTICATION_REQUIRED",
              "NON_AUTHORITATIVE_INFORMATION",
              "NOT_ACCEPTABLE",
              "NOT_EXTENDED",
              "NOT_FOUND",
              "NOT_IMPLEMENTED",
              "NOT_MODIFIED",
              "NO_CONTENT",
              "OK",
              "PARTIAL_CONTENT",
              "PAYLOAD_TOO_LARGE",
              "PAYMENT_REQUIRED",
              "PERMANENT_REDIRECT",
              "PRECONDITION_FAILED",
              "PRECONDITION_REQUIRED",
              "PROCESSING",
              "PROXY_AUTHENTICATION_REQUIRED",
              "REQUESTED_RANGE_NOT_SATISFIABLE",
              "REQUEST_ENTITY_TOO_LARGE",
              "REQUEST_HEADER_FIELDS_TOO_LARGE",
              "REQUEST_TIMEOUT",
              "REQUEST_URI_TOO_LONG",
              "RESET_CONTENT",
              "SEE_OTHER",
              "SERVICE_UNAVAILABLE",
              "SWITCHING_PROTOCOLS",
              "TEMPORARY_REDIRECT",
              "TOO_EARLY",
              "TOO_MANY_REQUESTS",
              "UNAUTHORIZED",
              "UNAVAILABLE_FOR_LEGAL_REASONS",
              "UNPROCESSABLE_ENTITY",
              "UNSUPPORTED_MEDIA_TYPE",
              "UPGRADE_REQUIRED",
              "URI_TOO_LONG",
              "USE_PROXY",
              "VARIANT_ALSO_NEGOTIATES"
            ]
          },
          "view": {
            "$ref": "#/components/schemas/View"
          },
          "viewName": {
            "type": "string"
          }
        }
      },
      "ResponseEntity": {
        "title": "ResponseEntity",
        "type": "object",
        "properties": {
          "body": {
            "type": "object"
          },
          "statusCode": {
            "type": "string",
            "enum": [
              "ACCEPTED",
              "ALREADY_REPORTED",
              "BAD_GATEWAY",
              "BAD_REQUEST",
              "BANDWIDTH_LIMIT_EXCEEDED",
              "CHECKPOINT",
              "CONFLICT",
              "CONTINUE",
              "CREATED",
              "DESTINATION_LOCKED",
              "EXPECTATION_FAILED",
              "FAILED_DEPENDENCY",
              "FORBIDDEN",
              "FOUND",
              "GATEWAY_TIMEOUT",
              "GONE",
              "HTTP_VERSION_NOT_SUPPORTED",
              "IM_USED",
              "INSUFFICIENT_SPACE_ON_RESOURCE",
              "INSUFFICIENT_STORAGE",
              "INTERNAL_SERVER_ERROR",
              "I_AM_A_TEAPOT",
              "LENGTH_REQUIRED",
              "LOCKED",
              "LOOP_DETECTED",
              "METHOD_FAILURE",
              "METHOD_NOT_ALLOWED",
              "MOVED_PERMANENTLY",
              "MOVED_TEMPORARILY",
              "MULTIPLE_CHOICES",
              "MULTI_STATUS",
              "NETWORK_AUTHENTICATION_REQUIRED",
              "NON_AUTHORITATIVE_INFORMATION",
              "NOT_ACCEPTABLE",
              "NOT_EXTENDED",
              "NOT_FOUND",
              "NOT_IMPLEMENTED",
              "NOT_MODIFIED",
              "NO_CONTENT",
              "OK",
              "PARTIAL_CONTENT",
              "PAYLOAD_TOO_LARGE",
              "PAYMENT_REQUIRED",
              "PERMANENT_REDIRECT",
              "PRECONDITION_FAILED",
              "PRECONDITION_REQUIRED",
              "PROCESSING",
              "PROXY_AUTHENTICATION_REQUIRED",
              "REQUESTED_RANGE_NOT_SATISFIABLE",
              "REQUEST_ENTITY_TOO_LARGE",
              "REQUEST_HEADER_FIELDS_TOO_LARGE",
              "REQUEST_TIMEOUT",
              "REQUEST_URI_TOO_LONG",
              "RESET_CONTENT",
              "SEE_OTHER",
              "SERVICE_UNAVAILABLE",
              "SWITCHING_PROTOCOLS",
              "TEMPORARY_REDIRECT",
              "TOO_EARLY",
              "TOO_MANY_REQUESTS",
              "UNAUTHORIZED",
              "UNAVAILABLE_FOR_LEGAL_REASONS",
              "UNPROCESSABLE_ENTITY",
              "UNSUPPORTED_MEDIA_TYPE",
              "UPGRADE_REQUIRED",
              "URI_TOO_LONG",
              "USE_PROXY",
              "VARIANT_ALSO_NEGOTIATES"
            ]
          },
          "statusCodeValue": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "TransferDto": {
        "title": "TransferDto",
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "format": "double"
          },
          "fromIban": {
            "type": "string"
          },
          "toIban": {
            "type": "string"
          }
        }
      },
      "View": {
        "title": "View",
        "type": "object",
        "properties": {
          "contentType": {
            "type": "string"
          }
        }
      }
    }
  }
}